Move Create New Word button
Plus some refactoring
